+++ pygccxml/declarations/declaration.py
@@ -335,5 +335,6 @@ class declaration_t(object):
         """
         warnings.warn(
             "The i_depend_on_them() method is deprecated.\n" +
-            "Please use the declarations.i_depend_on_them() method instead.\n",
+            "Please use the declarations.get_dependencies_from_decl()"  +
+            "method instead.\n",
             DeprecationWarning)